Add new commands to fix npm run setup inside a project
