+++ packages/strapi/lib/Strapi.js
@@ -176,6 +176,8 @@ class Strapi extends EventEmitter {
   }
 
   async load() {
+    await this.enhancer();
+
     this.app.use(async (ctx, next) => {
       if (ctx.request.url === '/_health' && ctx.request.method === 'HEAD') {
         ctx.set('strapi', 'You are so French!');
@@ -213,6 +215,7 @@ class Strapi extends EventEmitter {
 
     // Harmonize plugins configuration.
     await plugins.call(this);
+    console.log('INITIALISED ENDED');
   }
 
   reload() {
